fix debug configurations issue
