+++ uisdk/src/main/java/tech/dojo/pay/uisdk/presentation/PaymentFlowContainerActivity.kt
@@ -29,7 +29,6 @@ import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
 import tech.dojo.pay.sdk.DojoPaymentResult
 import tech.dojo.pay.sdk.DojoSdk
-import tech.dojo.pay.sdk.card.entities.DojoSDKDebugConfig
 import tech.dojo.pay.sdk.card.presentation.card.handler.DojoCardPaymentHandler
 import tech.dojo.pay.sdk.card.presentation.card.handler.DojoSavedCardPaymentHandler
 import tech.dojo.pay.sdk.card.presentation.card.handler.DojoVirtualTerminalHandler
@@ -81,10 +80,9 @@ class PaymentFlowContainerActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        configureDojoPayCore()
         disableScreenRecord()
         lockToPortrait()
-        configureDojoSDKDebugConfig()
-        configureDojoPayCore()
         setContent {
             DojoTheme {
                 val forceLightMode = DojoSDKDropInUI.dojoThemeSettings?.forceLightMode ?: false
@@ -134,7 +132,6 @@ class PaymentFlowContainerActivity : AppCompatActivity() {
     }
 
     private fun configureDojoPayCore() {
-        configureDojoSDKDebugConfig()
         gpayPaymentHandler = DojoSdk.createGPayHandler(this) {
             paymentFlowViewModel.updateGpayPaymentState(false)
             paymentFlowViewModel.navigateToPaymentResult(it)
@@ -150,18 +147,6 @@ class PaymentFlowContainerActivity : AppCompatActivity() {
         virtualTerminalHandler = DojoSdk.createVirtualTerminalPaymentHandler(this) {
             paymentFlowViewModel.updatePaymentState(false)
             paymentFlowViewModel.navigateToPaymentResult(it)
-        }
-    }
-
-    private fun configureDojoSDKDebugConfig() {
-        if (DojoSDKDropInUI.dojoSDKDebugConfig != null) {
-            DojoSDKDropInUI.dojoSDKDebugConfig?.let { DojoSdk.dojoSDKDebugConfig = it }
-        } else {
-            val dojoSDKDebugConfig = DojoSDKDebugConfig(
-                isSandboxWallet = paymentFlowViewModel.isPaymentInSandBoxEnvironment(),
-                isSandboxIntent = paymentFlowViewModel.isPaymentInSandBoxEnvironment(),
-            )
-            DojoSdk.dojoSDKDebugConfig = dojoSDKDebugConfig
         }
     }
